Added spectral duplicate detection
The hypergeometric distribution is used to compute a probability that each pair of spectra would match this many deconvolved masses or more by random chance.  The Redundancy metric is now exported per RAW file, expressing the percentage of all MSn-MSn matches within the RAW that produce a negative log probability of 100 or greater.