Shift to Visual Studio, Export mzQC format
I have shifted the source to Visual Studio Community in order to incorporate @mwalzer's mzQC library prototype.  This is the first version of the source that outputs mzQC format along with its TSV reports.  mzQC format is not finalized because I need some terms added to the CV.  TDAuditor now detects and reports the deconvolution engine that produced the supplied msAlign or MGF files.
At the moment, TDAuditor-GUI is not included in the build process.